+++ NEWS.md
@@ -3,6 +3,7 @@
 * `query()` can now batch annotate several images with the same prompt by passing `images` as a list (one element per query)
 * updated image annotation vignette to explain batch annotation
 * fixes issue in tests
+* bug fixes
 
 # rollama 0.3.0